add to gitignore venv, refactor utils
